requirements.txt
python-telegram-bot
google-generativeai
SpeechRecognition
PyAudio
wget
pydub
ffmpeg-python